Basic-Minimap-Functionality
Code added to change view from close up of snail to view of full map. To
make code work, the following changes need to be made to the scene:
- Two Layers need to be added: Background and Minimap. basicMap and
Escargo_Background_1 objects should be in the Background layer. Anything
that you wish to be displayed as a part of the minimap should be marked
as a part of the minimap layer (i.e. symbols showing slime and snail
positions rather than the players themselves).
- Minimap keys will have to be assigned for players other than player 1.
Default is "Q".