Snailling AI /kinda/ works...and other adventures in crashing Unity repeatedly
Snailling AI /kinda/ works. The ai itself seems to be okay. The issue is
with running the function too many times. The idea now is that it runs a
pathfind to the player when the snailling first appears. Then each time
the player set slime pathfinding is run on the head snaillings to the
player, then from the 2nd snailling to the head snailling, on down the
line...it just does weird shit. Also trying to fix collision issues with
ignorecollisionlayer but it's not working.